fix for Xcode 6.3 warning

+++ cocos2d/CCTextureAtlas.m
@@ -257,7 +257,7 @@
 	if( oldIndex == newIndex )
 		return;
 
-	NSUInteger howMany = labs( oldIndex - newIndex);
+	NSUInteger howMany = labs( (NSInteger)oldIndex - (NSInteger)newIndex);
 	NSUInteger dst = oldIndex;
 	NSUInteger src = oldIndex + 1;
 	if( oldIndex > newIndex) {
